style: renamed folders to start with small case letters
