--- src/Config/mapConfig.js
@@ -1,9 +1,0 @@
-const MAP_CENTER = [34.78, 32.07];
-const DEFAULT_PROJECTION = 'EPSG:4326';
-const DEFAULT_ZOOM = 5;
-
-module.exports = {
-  MAP_CENTER,
-  DEFAULT_PROJECTION,
-  DEFAULT_ZOOM,
-};
