--- src/Store/reducers.js
@@ -1,5 +1,0 @@
-import dataReducer from './Reducers/dataReducer';
-
-export default {
-  data: dataReducer,
-};
