+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import { Provider } from 'react-redux';
-import store from './Store/store';
+import store from './store/store';
 import { SnackbarProvider } from 'notistack';
 
 ReactDOM.render(
