+++ src/store/reducers.js
@@ -1,0 +1,5 @@
+import dataReducer from './reducers/dataReducer';
+
+export default {
+  data: dataReducer,
+};
